requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pypdf
pydantic>=2.9,<3.0
pydantic-settings
langchain-openai
langchain-google-genai
langchain-community
openai
crewai
crewai-tools
sqlalchemy==2.0.36